requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cloudinary==1.41.0
cryptography==43.0.3
Django==5.1.2
djangorestframework==3.15.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
